% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bca.R
\name{bayes_lin_reg_stats}
\alias{bayes_lin_reg_stats}
\title{BCA linear regression summary stats}
\usage{
bayes_lin_reg_stats(
  df,
  outcome,
  covariates,
  weights = NULL,
  k = 0,
  n_sites = 0
)
}
\arguments{
\item{df}{Dataframe containing covariates X and outcome y.}

\item{outcome}{Name of outcome y.}

\item{covariates}{Vector of covariate names.}

\item{weights}{Vector of weights for weighted linear regression. Default
NULL.}

\item{k}{Site number in case local intercepts are used. Default 0 corresponds
to not using local intercepts.}

\item{n_sites}{Number of sites. Default 0.}
}
\value{
List with transmitted summary statistics for each local site.
}
\description{
Takes in dataframe with (X, y) and returns sufficient statistics
  (X'X, X'y, y'y, n)
}
\author{
Peter Degen
}
