% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaySeq.R
\name{bayes_local_glm}
\alias{bayes_local_glm}
\title{Local params}
\usage{
bayes_local_glm(bstats, center_identity, family, alpha = 0.05)
}
\arguments{
\item{bstats}{List of summary statistics from each local site, e.g. obbject
returned by `bca_iterate_sites`.}

\item{center_identity}{The id of the site to be removed.}

\item{family}{Family to pass to glm() call, e.g. "gaussian".}

\item{alpha}{Numeric. Significance level for credible intervals. Default
0.05.}
}
\value{
List with beta hat and credible intervals.
}
\description{
Compute local glm parameters from summary statistics.
}
\author{
Peter Degen
}
