% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bca.R
\name{bca_glm_oneshot}
\alias{bca_glm_oneshot}
\title{BCA GLM one-shot}
\usage{
bca_glm_oneshot(sumstats, pooling = "none", glm_prior_lamda = 0, alpha = 0.05)
}
\description{
Perform one-shot computation of GLM parameter estimates given
  transmitted summary statistics. Users should use `bca_oneshot`, which wraps
  this method.
}
