% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bca.R
\name{bca_iterate_sites}
\alias{bca_iterate_sites}
\title{BCA iterate over local sites}
\usage{
bca_iterate_sites(
  model,
  family,
  data_split,
  use_local_intercepts,
  outcome = NULL,
  covariates = NULL,
  alpha = 0.05
)
}
\arguments{
\item{model}{Formula object relating outcome to covariates.}

\item{family}{Family to pass to glm() call, e.g. "gaussian".}

\item{data_split}{List, each element containing a dataframe with the IPD
from a local site.}

\item{use_local_intercepts}{Logical. If true, use fixed site-specific
intercepts for each local site.}

\item{outcome}{Name of outcome y.}

\item{covariates}{Vector of covariate names.}

\item{alpha}{Numeric. Significance level for credible intervals. Default
0.05.}

\item{center_name}{Character (optional). Name of covariate denoting site
identity. Must be supplied if `use_local_intercepts` is TRUE.}
}
\value{
List with transmitted summary statistics for each local site.
}
\description{
Pre-processing wrapper for testing BCA methods using individual participant data.
}
\author{
Peter Degen
}
