% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bca.R
\name{bca_oneshot}
\alias{bca_oneshot}
\title{BCA one-shot}
\usage{
bca_oneshot(
  sumstats,
  family,
  use_local_intercepts,
  use_local_variances = FALSE,
  epsilon = 1e-10,
  alpha = 0.05,
  covariates = NULL,
  glm_prior_lamda = 0
)
}
\arguments{
\item{sumstats}{List of summary statistics from each local site, e.g. obbject
returned by `bca_iterate_sites`.}

\item{family}{Family to pass to glm() call, e.g. "gaussian".}

\item{use_local_intercepts}{Logical. If true, use fixed site-specific
intercepts for each local site.}

\item{use_local_variances}{Logical. If true, use fixed site-specific residual
variances for each local site. Default FALSE.}

\item{epsilon}{Numeric. Regularization for prior hyperparameters. Default:
1e-10.}

\item{alpha}{Numeric. Significance level for credible intervals. Default:
0.05.}

\item{glm_prior_lamda}{Numeric (optional). Prior for diagonal precision
matrix in normal model, known variance case. See also Equation 7 in Jonker,
Pazira and Coolen (2024). Default: 0.}

\item{center_name}{Character (optional). Name of covariate denoting site
identity. Must be supplied if `family` is `gaussian`.}
}
\value{
List of oneshot parameter estimates
}
\description{
Perform one-shot computation of parameter estimates given
  transmitted summary statistics.
}
\author{
Peter Degen
}
