% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaySeq.R
\name{box_check_all_sites}
\alias{box_check_all_sites}
\title{Box prior predictive tail probabilities for all sites}
\usage{
box_check_all_sites(
  data_split,
  params_oneshot,
  bstats,
  family,
  local_sigma2s,
  covariates,
  n_sites,
  use_local_intercepts,
  center_name
)
}
\arguments{
\item{data_split}{The data split by local sites.}

\item{params_oneshot}{Parameters returned by `bca_oneshot`.}

\item{bstats}{List of summary statistics from each local site, e.g. obbject
returned by `bca_iterate_sites`.}

\item{family}{Family to pass to glm() call, e.g. "gaussian".}

\item{local_sigma2s}{Numeric vector with local residual variances from glm().}

\item{covariates}{Vector of covariate names.}

\item{n_sites}{Number of sites.}

\item{use_local_intercepts}{Logical. If true, use fixed site-specific
intercepts for each local site.}

\item{center_name}{Character. Name of covariate denoting site identity.}
}
\value{
List with pboxes and reduced parameters for all sites.
}
\description{
Calculate p_Box for all sites.
}
\author{
Peter Degen
}
