% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{forest_box}
\alias{forest_box}
\title{Multivariate forest plot wrapper}
\usage{
forest_box(
  params_oneshot,
  sumstats,
  family,
  pboxes = NULL,
  use_local_intercepts,
  center_name = "site",
  remove_intercept = FALSE,
  alpha = 0.05,
  remove_sigma2 = TRUE,
  ...
)
}
\arguments{
\item{params_oneshot}{Parameters returned by `bca_oneshot`.}

\item{sumstats}{List of summary statistics from each local site, e.g. obbject
returned by `bca_iterate_sites`.}

\item{family}{Family to pass to glm() call, e.g. "gaussian".}

\item{use_local_intercepts}{Logical. If true, use fixed site-specific
intercepts for each local site.}

\item{center_name}{Character. Name of covariate denoting site identity.}

\item{alpha}{Numeric. Significance level for credible intervals. Default
0.05.}

\item{...}{Further arguments passed to `prepare_forest_plot`.}
}
\value{
ggplot2 object.
}
\description{
Wrapper that computes pbox and makes forest plot in one call.
}
\author{
Peter Degen
}
