% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{forest_plot}
\alias{forest_plot}
\title{Multivariate forest plot}
\usage{
forest_plot(
  df_forest,
  pboxes = NULL,
  outfile = NULL,
  alpha = 0.05,
  nrow = 1,
  inline_plot = FALSE,
  use_log_scale = TRUE,
  order_box = TRUE
)
}
\arguments{
\item{df_forest}{Dataframe returned by `prepare_forest_plot`.}

\item{pboxes}{Optional numeric vector of Box p-values, contained in list
returned by `box_check_all_sites`.}

\item{outfile}{Optional character. If not null, save figure in this file.}

\item{alpha}{Numeric. Significance level for credible intervals. Default
0.05.}

\item{nrow}{Numeric. (default: 1)}

\item{inline_plot}{Logical. If TRUE, adjust width and height of inline plot
in Jupyter notebooks using `options()`. (default: FALSE)}

\item{use_log_scale}{Logical. If TRUE, print pbox as an S-value, i.e. -log2(pbox) (default: FALSE)}

\item{order_box}{FALSE Logical. If TRUE, order sites by p-Box. (default: FALSE)}
}
\value{
ggplot2 object.
}
\description{
Creates a forest plot with a facet for each covariate.
  Optionally prints Box's prior predictive tail probability for each site.
}
\author{
Peter Degen
}
