% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bca.R
\name{get_reduced_params}
\alias{get_reduced_params}
\title{Get reduced params}
\usage{
get_reduced_params(center_identity, params_oneshot, sumstats, family)
}
\arguments{
\item{center_identity}{The id of the site to be removed.}

\item{params_oneshot}{Parameters returned by bca_oneshot.}

\item{sumstats}{List of summary statistics from each local site, e.g. obbject
returned by `bca_iterate_sites`.}

\item{family}{Family to pass to glm() call, e.g. "gaussian".}
}
\value{
List with reduced parameter estimates.
}
\description{
Use Reverse-Bayes to get reduced parameters obtained by leaving
  out one local site from the full posterior.
}
\author{
Peter Degen
}
