% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bca.R
\name{pred_check}
\alias{pred_check}
\title{Box prior predictive tail probabilities for all sites}
\usage{
pred_check(
  params_oneshot,
  sumstats,
  family,
  use_local_intercepts,
  center_name = NULL,
  remove_intercept = FALSE
)
}
\arguments{
\item{params_oneshot}{Parameters returned by `bca_oneshot`.}

\item{sumstats}{List of summary statistics from each local site, e.g. obbject
returned by `bca_iterate_sites`.}

\item{family}{Family to pass to glm() call, e.g. "gaussian".}

\item{use_local_intercepts}{Logical. If true, use fixed site-specific
intercepts for each local site.}

\item{center_name}{Character. Name of covariate denoting site identity.}

\item{remove_intercept}{Logical. If true, ignore intercept in pbox
calculation. Default: FALSE.}

\item{data_split}{The data split by local sites.}

\item{covariates}{Vector of covariate names.}

\item{n_sites}{Number of sites.}
}
\value{
List with pboxes and reduced parameters for all sites.
}
\description{
Calculate p_Box for all sites.
}
\author{
Peter Degen
}
