% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{prepare_forest_plot}
\alias{prepare_forest_plot}
\title{Prepare forest plot}
\usage{
prepare_forest_plot(df_bca, sumstats, family, alpha = 0.05)
}
\arguments{
\item{df_bca}{Dataframe with BCA estimates returned by `tidy_results`.}

\item{sumstats}{List of summary statistics from each local site, e.g. obbject
returned by `bca_iterate_sites`.}

\item{alpha}{Numeric. Significance level for credible intervals. Default
0.05.}
}
\value{
Dataframe with parameter estimates and confidence intervals for all
  sites and covariates.
}
\description{
Get clean dataframe with local glm parameters from summary statistics for all
  sites plus the federated BCA estimates.
}
\author{
Peter Degen
}
