% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaySeq.R
\name{tidy_results}
\alias{tidy_results}
\title{BCA tidy results}
\usage{
tidy_results(params_oneshot, use_local_intercepts)
}
\arguments{
\item{params_oneshot}{Parameters returned by `bca_oneshot`.}

\item{use_local_intercepts}{Logical. If true, use fixed site-specific
intercepts for each local site.}
}
\value{
Dataframe
}
\description{
Tidy results from one-shot
}
\author{
Peter Degen
}
