% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transform_X}
\alias{transform_X}
\title{transform X}
\usage{
transform_X(X)
}
\arguments{
\item{x}{Design matrix X with covariates}
}
\value{
Transformed design matrix X
}
\description{
Take data vector X and convert factors with numeric. Binary
  factors will be encoded by {0, 1}. Factors with more than two levels will
  be transformed using one-hot encoding, dropping one level to avoud
  multicollinearity.
}
\author{
Peter Degen
}
