% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaySeq.R
\name{bca_iterate_sites}
\alias{bca_iterate_sites}
\title{BCA iterate over local sites}
\usage{
bca_iterate_sites(
  outcome,
  covariates,
  model,
  family,
  data_split,
  use_local_intercepts,
  center_name = NULL
)
}
\arguments{
\item{outcome}{Name of outcome y.}

\item{covariates}{Vector of covariate names.}

\item{model}{Formula object relating outcome to covariates.}

\item{family}{Family to pass to glm() call, e.g. "gaussian".}

\item{data_split}{List, each element containing a dataframe with the data.
from a local site.}

\item{use_local_intercepts}{Logical. If true, use fixed site-specific
intercepts for each local site.}

\item{center_name}{Character (optional). Name of covariate denoting site identity. Must
be supplied if `use_local_intercepts` is TRUE.}
}
\value{
List with transmitted summary statistics for each local site.
}
\description{
Pre-processing wrapper for BCA methods.
}
\author{
Peter Degen
}
