% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaySeq.R
\name{bca_iterate_sites}
\alias{bca_iterate_sites}
\title{BCA iterate over local sites}
\usage{
bca_iterate_sites(
  outcome,
  covariates,
  Method,
  data_split,
  use_local_intercepts,
  center_name = NULL
)
}
\arguments{
\item{outcome}{Outcome vector y}
}
\value{
List over all local sites with transmitted summary statistics
}
\description{
Pre-processing wrapper for BCA methods.
}
\author{
Peter Degen
}
