% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaySeq.R
\name{bca_oneshot}
\alias{bca_oneshot}
\title{BCA one-shot}
\usage{
bca_oneshot(
  bstats,
  n_sites,
  use_local_intercepts,
  use_local_variances,
  family,
  covariates,
  epsilon = 1e-10,
  center_name = NULL,
  CI = "normal",
  return_post_params = FALSE,
  alpha = 0.05
)
}
\arguments{
\item{bstats}{List of summary statistics returned by bca_iterate_sites}
}
\value{
List of oneshot parameter estimates
}
\description{
Perform one-shot computation
}
\author{
Peter Degen
}
