% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bca.R
\name{get_pred_prob}
\alias{get_pred_prob}
\title{Box prior predictive tail probability}
\usage{
get_pred_prob(
  center_identity,
  sumstats,
  n_sites,
  reduced_params,
  use_local_intercepts,
  remove_intercept = FALSE
)
}
\arguments{
\item{center_identity}{The id of the site to be checked}

\item{sumstats}{List of summary statistics from each local site, e.g. obbject
returned by `bca_iterate_sites`.}

\item{n_sites}{Number of sites.}

\item{reduced_params}{Reduced posterior parameters returnd by
`get_reduced_params`.}

\item{use_local_intercepts}{Logical. If true, use fixed site-specific
intercepts for each local site.}

\item{remove_intercept}{Logical. If true, ignore intercept in pbox
calculation. Default: FALSE.}

\item{covariates}{Vector of covariate names.}
}
\value{
The Box prior predictive tail probability.
}
\description{
Check if estimate from specific site is compatible with
  posterior from all other centers.
}
\author{
Peter Degen
}
